small changes made to DDPG algo
