+++ benchmarking.py
@@ -86,9 +86,9 @@ def main(INIT_HP, MUTATION_PARAMS, NET_CONFIG):
 
 if __name__ == '__main__':
     INIT_HP = {
-        'ENV_NAME': 'LunarLanderContinuous-v2',   # Gym environment name
-        'ALGO': 'TD3',                  # Algorithm
-        #'DOUBLE': True,                 # Use double Q-learning
+        'ENV_NAME': 'LunarLander-v2',   # Gym environment name
+        'ALGO': 'DQN',                  # Algorithm
+        'DOUBLE': True,                 # Use double Q-learning
         # Swap image channels dimension from last to first [H, W, C] -> [C, H, W]
         'CHANNELS_LAST': False,
         'BATCH_SIZE': 256,              # Batch size
